+++ embassy-stm32/src/lib.rs
@@ -151,7 +151,7 @@ pub use crate::_generated::interrupt;
 /// Macro to bind interrupts to handlers.
 ///
 /// This defines the right interrupt handlers, and creates a unit struct (like `struct Irqs;`)
-/// and implements the right [`Binding`]s for it. You can pass this struct to drivers to
+/// and implements the right [`Binding`](crate::interrupt::typelevel::Binding)s for it. You can pass this struct to drivers to
 /// prove at compile-time that the right interrupts have been bound.
 ///
 /// Example of how to bind one interrupt:
@@ -194,6 +194,19 @@ macro_rules! bind_interrupts {
         #[derive(Copy, Clone)]
         $(#[$outer])*
         $vis struct $name;
+
+        impl $name {
+            #[doc = r"Convenience method to call Binding::into_any(). Unlike the trait method, can be called with a turbofish."]
+            pub fn as_any<
+                I: $crate::interrupt::typelevel::Interrupt,
+                H: $crate::interrupt::typelevel::Handler<I>
+            >() -> $crate::interrupt::typelevel::AnyBinding
+            where
+                Self: $crate::interrupt::typelevel::Binding<I, H>
+            {
+                <Self as $crate::interrupt::typelevel::Binding<I, H>>::into_any()
+            }
+        }
 
         $(
             #[allow(non_snake_case)]