+++ examples/boot/application/stm32wl/src/bin/a.rs
@@ -9,9 +9,11 @@ use embassy_boot_stm32::{AlignedBuffer, FirmwareUpdater, FirmwareUpdaterConfig};
 use embassy_embedded_hal::adapter::BlockingAsync;
 use embassy_executor::Spawner;
 use embassy_stm32::SharedData;
-use embassy_stm32::exti::ExtiInput;
+use embassy_stm32::bind_interrupts;
+use embassy_stm32::exti::{self, ExtiInput};
 use embassy_stm32::flash::{Flash, WRITE_SIZE};
 use embassy_stm32::gpio::{Level, Output, Pull, Speed};
+use embassy_stm32::interrupt;
 use embassy_sync::mutex::Mutex;
 use panic_reset as _;
 
@@ -19,6 +21,11 @@ use panic_reset as _;
 static APP_B: &[u8] = &[0, 1, 2, 3];
 #[cfg(not(feature = "skip-include"))]
 static APP_B: &[u8] = include_bytes!("../../b.bin");
+
+bind_interrupts!(
+    pub struct Irqs{
+        EXTEXTI0I2_3 => exti::InterruptHandler<interrupt::typelevel::EXTI0>;
+});
 
 #[unsafe(link_section = ".shared_data")]
 static SHARED_DATA: MaybeUninit<SharedData> = MaybeUninit::uninit();
@@ -29,7 +36,12 @@ async fn main(_spawner: Spawner) {
     let flash = Flash::new_blocking(p.FLASH);
     let flash = Mutex::new(BlockingAsync::new(flash));
 
-    let mut button = ExtiInput::new(p.PA0, p.EXTI0, Pull::Up);
+    let mut button = ExtiInput::new(
+        p.PA0,
+        p.EXTI0,
+        Pull::Up,
+        Irqs::as_any::<interrupt::typelevel::EXTI0, exti::InterruptHandler<interrupt::typelevel::EXTI0>>(),
+    );
 
     let mut led = Output::new(p.PB9, Level::Low, Speed::Low);
     led.set_high();